+++ supercell/__init__.py
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 #
-__version__ = (0, 6, 0, 'alpha')
+__version__ = (0, 6, 0)